+++ miner/src/create_block_template/mod.rs
@@ -7,3 +7,6 @@ pub mod new_header_service;
 //mod test_create_block_template;
 pub mod block_builder_service;
 pub mod state_check;
+
+#[cfg(test)]
+mod test_state_check;
